Move tracing to common module
