--- packages/server/customer-os-api/tracing/consts.go
@@ -1,7 +1,0 @@
-package tracing
-
-const (
-	SpanTagTenant    = "tenant"
-	SpanTagUserId    = "user-id"
-	SpanTagComponent = "component"
-)
