+++ packages/server/customer-os-api/tracing/utils.go
@@ -2,15 +2,19 @@ package tracing
 
 import (
 	"context"
-	"encoding/json"
 	"github.com/99designs/gqlgen/graphql"
-	"github.com/labstack/echo/v4"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/common"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/constants"
+	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/tracing"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/ext"
-	"google.golang.org/grpc/metadata"
 	"net/http"
+)
+
+const (
+	SpanTagTenant    = "tenant"
+	SpanTagUserId    = "user-id"
+	SpanTagComponent = "component"
 )
 
 func StartGraphQLTracerSpan(ctx context.Context, operationName string, operationContext *graphql.OperationContext) (context.Context, opentracing.Span) {
@@ -39,83 +43,8 @@ func StartHttpServerTracerSpanWithHeader(ctx context.Context, operationName stri
 	return opentracing.ContextWithSpan(ctx, serverSpan), serverSpan
 }
 
-func StartHttpServerTracerSpan(c echo.Context, operationName string) (context.Context, opentracing.Span) {
-	spanCtx, err := opentracing.GlobalTracer().Extract(opentracing.HTTPHeaders, opentracing.HTTPHeadersCarrier(c.Request().Header))
-
-	if err != nil {
-		serverSpan := opentracing.GlobalTracer().StartSpan(operationName)
-		return opentracing.ContextWithSpan(c.Request().Context(), serverSpan), serverSpan
-	}
-
-	serverSpan := opentracing.GlobalTracer().StartSpan(operationName, ext.RPCServerOption(spanCtx))
-	return opentracing.ContextWithSpan(c.Request().Context(), serverSpan), serverSpan
-}
-
-func StartGrpcServerTracerSpan(ctx context.Context, operationName string) (context.Context, opentracing.Span) {
-	textMapCarrierFromMetaData := GetTextMapCarrierFromMetaData(ctx)
-
-	span, err := opentracing.GlobalTracer().Extract(opentracing.TextMap, textMapCarrierFromMetaData)
-	if err != nil {
-		serverSpan := opentracing.GlobalTracer().StartSpan(operationName)
-		ctx = opentracing.ContextWithSpan(ctx, serverSpan)
-		return ctx, serverSpan
-	}
-
-	serverSpan := opentracing.GlobalTracer().StartSpan(operationName, ext.RPCServerOption(span))
-	ctx = opentracing.ContextWithSpan(ctx, serverSpan)
-	return ctx, serverSpan
-}
-
-func GetTextMapCarrierFromMetaData(ctx context.Context) opentracing.TextMapCarrier {
-	metadataMap := make(opentracing.TextMapCarrier)
-	if md, ok := metadata.FromIncomingContext(ctx); ok {
-		for key := range md.Copy() {
-			metadataMap.Set(key, md.Get(key)[0])
-		}
-	}
-	return metadataMap
-}
-
-func InjectTextMapCarrier(spanCtx opentracing.SpanContext) (opentracing.TextMapCarrier, error) {
-	m := make(opentracing.TextMapCarrier)
-	if err := opentracing.GlobalTracer().Inject(spanCtx, opentracing.TextMap, m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func ExtractTextMapCarrier(spanCtx opentracing.SpanContext) opentracing.TextMapCarrier {
-	textMapCarrier, err := InjectTextMapCarrier(spanCtx)
-	if err != nil {
-		return make(opentracing.TextMapCarrier)
-	}
-	return textMapCarrier
-}
-
-func ExtractTextMapCarrierBytes(spanCtx opentracing.SpanContext) []byte {
-	textMapCarrier, err := InjectTextMapCarrier(spanCtx)
-	if err != nil {
-		return []byte("")
-	}
-
-	dataBytes, err := json.Marshal(&textMapCarrier)
-	if err != nil {
-		return []byte("")
-	}
-	return dataBytes
-}
-
-func InjectTextMapCarrierToGrpcMetaData(ctx context.Context, spanCtx opentracing.SpanContext) context.Context {
-	if textMapCarrier, err := InjectTextMapCarrier(spanCtx); err == nil {
-		md := metadata.New(textMapCarrier)
-		ctx = metadata.NewOutgoingContext(ctx, md)
-	}
-	return ctx
-}
-
 func TraceErr(span opentracing.Span, err error) {
-	span.SetTag("error", true)
-	span.LogKV("error_msg", err.Error())
+	tracing.TraceErr(span, err)
 }
 
 func setDefaultSpanTags(ctx context.Context, span opentracing.Span) {
